requirements.txt
# Web scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.4.0

# Data handling
pandas>=1.5.0

# Email handling
# smtplib (part of the Python standard library)

# Testing
pytest>=7.0.0

# Web driver management for Selenium
webdriver-manager>=3.8.0

# HTML parsing
lxml>=4.9.0

# Date/time handling
python-dateutil>=2.8.0

# Environment variables
python-dotenv>=0.20.0
